inference/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
transformers>=4.51.0
torch>=2.1.0
google-cloud-storage>=2.10.0
flash-attn>=2.3.0
sentencepiece>=0.1.99

